加入歌词界面

+++ app/src/main/java/com/android/wy/news/music/lrc/Lrc.kt
@@ -1,6 +1,6 @@
 package com.android.wy.news.music.lrc
 
 data class Lrc(
-    var time: Long,
+    var time: Float,
     var text: String
 )
