+++ app/src/main/java/com/android/wy/news/music/lrc/LrcView.kt
@@ -577,7 +577,7 @@ class LrcView : View {
     }
 
     interface OnPlayIndicatorLineListener {
-        fun onPlay(time: Long, content: String?)
+        fun onPlay(time: Float, content: String?)
     }
